refactor(script): split map buildins into buildin_map.c
Issue #61 first step: move warp/announce/mapflag/pvp-gvg helpers out of script.c and expose shared script APIs via script_internal.h.
