+++ vcproj/map-server.vcxproj
@@ -137,6 +137,7 @@
     <ClInclude Include="..\src\map\quest.h" />
     <ClInclude Include="..\src\map\ranking.h" />
     <ClInclude Include="..\src\map\script.h" />
+    <ClInclude Include="..\src\map\script_internal.h" />
     <ClInclude Include="..\src\map\skill.h" />
     <ClInclude Include="..\src\map\skill_internal.h" />
     <ClInclude Include="..\src\map\status.h" />
@@ -206,6 +207,7 @@
     <ClCompile Include="..\src\map\quest.c" />
     <ClCompile Include="..\src\map\ranking.c" />
     <ClCompile Include="..\src\map\script.c" />
+    <ClCompile Include="..\src\map\buildin_map.c" />
     <ClCompile Include="..\src\map\skill.c" />
     <ClCompile Include="..\src\map\skill_castend_damage.c" />
     <ClCompile Include="..\src\map\skill_castend_ground.c" />
